Implemented async buffering of music
