+++ src/assets/sound.go
@@ -1,0 +1,9 @@
+package assets
+
+import "github.com/faiface/beep"
+
+type sound interface {
+	decreaseVolume()
+	play(mixer *beep.Mixer)
+	stop()
+}
